# docs/swagger.yaml
definitions:
...
  dtos.MessagesResponses:
    properties:
      message:
        type: string
    type: object
...
info:
  contact:
    email: cs@payuoge.com
    name: API Support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /transactions/checkout:
    get:
      consumes:
      - application/json
      description: do Get All checkout
      produces:
      - application/json
      responses:
        "200":
          description: message responses
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: not authorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get All Checkout access process
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: do create a create checkout
      produces:
      - application/json
      responses:
        "200":
          description: message responses
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: not authorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: CreateCheckout access process
      tags:
      - transactions
  /transactions/checkout/{id}:
    delete:
      consumes:
      - application/json
      description: do delete id checkout
      parameters:
      - description: delete by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message responses
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: not authorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: delete id Checkout access process
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: do Get id checkout
      produces:
      - application/json
      responses:
        "200":
          description: message responses
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: not authorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get id Checkout access process
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: do update id checkout
      produces:
      - application/json
      responses:
        "200":
          description: message responses
          schema:
            $ref: '#/definitions/dtos.MessagesResponses'
        "400":
          description: not authorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: update id Checkout access process
      tags:
      - transactions
  /transactions/orders:
    get:
      consumes:
      - application/json
      description: do get all order
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Get All Order access process
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: do create a order
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Create Orders access process
      tags:
      - transactions
  /transactions/orders/{id}:
    delete:
      consumes:
      - application/json
      description: do delete id order
      parameters:
      - description: delete by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: delete id order access process
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: do get id order
      parameters:
      - description: get id order
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: Get id order access process
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: do update id order
      parameters:
      - description: update id order
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: update id order access process
      tags:
      - transactions
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"